Drop unnecessary `sleep` calls
Because the main loop uses `await` when fetching new events to process,
if there are no events the main thread simply pauses execution until
there is more IO to be done.

The sleep call is not necessary due to the async nature of the library,
and will just add an artificial delay to events, as well as an extra
wakeup.

Closes #29